add option --rebuild, and --overwrite. If yaml file already exist, then buildtest will not do anything. --rebuild will recreate yaml file if file exist and name it as different file. --overwrite will rewrite the content to command.yaml. The option --ebyaml and --sysyaml are renamed to --app and --systempkg along with short options.
